pyproject.toml
[tool.poetry]
name = "thumbnail_generation_api"
version = "0.4.0"
description = "Rest API to generate thumbnails"
authors = [
    "Konstantinos Platis <konstantinos.platis@epfl.ch>",
    "Harry Robert Taylor Anderson <harry.anderson@epfl.ch>",
    "Daniel Burger <daniel.burger@epfl.ch>"
]
packages = [{ include = "api" }]

[tool.poetry.dependencies]
python = "==3.10.*"
fastapi = { version = "^0.110.3", extras = ["standard"] }
gunicorn = "^21.2.0"
uvicorn = "^0.25.0"
PyJWT = "^2.8.0"
pytest = "^8.1.2"
neurom = "^3.2.5"
matplotlib = "^3.7.4"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
h5py = "^3.10.0"
numpy = "^1.24.4"
python-multipart = "^0.0.9"
pydantic = "^2.7.1"
httpx = "0.27.0"

[tool.poetry.dev-dependencies]
black = "^23.11.0"
pylint = "^3.0.2"
ruff = "^0.3.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.FORMAT]
max-line-length = 120
good-names = ["version_template"]
disable = ["R0903"]

[tool.pylint.MASTER]
extension-pkg-whitelist = ["pydantic"]

[tool.black]
line-length = 120
